fix: useJustAppContext can specify type to anything
So that it can do implicit stuff and expect only the type a particular component needs.
i.e. ISP.

Yes, if the actual instance does not have those gizmo it will fail in runtime.

But the flexibility is there.

If we want to be sure,
the package can create a gizmo and create a specific context for that?

Need to detail those out.